Use an uneven test distribution
This commit changes how the number of tests are passed to
the workers from an evenly fashion to an uneven one in order to
take into account that otherwise the last tests would have a
bigger input size and take more time and have a bigger
computational cost.

Nevertheless, if the uneven distribution would not be sustainable
(e.g., the distribution would leave too many or too few tests for
the last workers) an even one will be used instead.